solved a bug in Freeze PID mode
There was a bug that caused disfunction in the thermostat when in PID
freeze mode. This was linked to the old use of speeds measured in K/min,
whereas we have now moved to Ohm/min for both the thernostat and the
core. NOTE: this is yet to be made for the jacket.