Corrected k6220 config and K2182 config
Noticed that k6220 configuration was identical for the Core and the
Service Wheatstone Bridges, which should be not, given the use of Delta
mode on the first. Also tried back and forth the use of 1  or 5
NPLCycles on the k2182 that serves the Wheatstone bridge, still trying
to identify the best.